Refactor error handling for makeSuggestRequest
Instead of throwing exceptions, return an Either, then match against it
and complete the request appropriately. This allows us to explicitly
spell out what to do for unauthorized and not-found requests, which
previously relied on behaviour where an exception is explicitly thrown
or encountered when trying to invoke Option.get

This change has makeSuggestRequest use StatusCodes instead of its own
items for errors. The exceptions in exc/package are not matchable, and
we have no time to properly refactor those, so this leaked abstraction
will have to do for now